Do not give a fatal error if attempting to compile ActiveQt on Unix
Gracefully disable, instead of giving fatal error, if attempting to
compile ActiveQt on platforms where it is not supported.

This is helpful for the sake of having a uniform test procedure.
Without this, any test procedure needs to encode the logic that it
shouldn't attempt to compile ActiveQt on platforms other than Windows.
It is cleaner if the ActiveQt project itself gracefully degrades to
"do nothing" on those platforms where it is not supported.

Reviewed-by: Jason McDonald
Change-Id: Id2544b4cbba604a5f76ba56a12bb801cb36e1646