WIP: Don't alias registered types to void *
Some types were being registered in multiple places in different ways.
This patch tries to ensure that types are preferentially registered in
the canonical qRegisterMetaType<T>(#T) form.

Where this is not possible, we first check whether the type is already
registered, and otherwise attempt aliasing to void*.

WIP: This is going to need some Q_DECLARE_OPAQUE_POINTER and I'll have
to fix the existing monsterifications... Anyone know if that macro is
already available in sub-module CIs?

Change-Id: I564b6ad3cd14daed9452d85a775c980f5772c703