Don't register previously registered types as aliases
Some types were being registered in multiple places in different ways.
This patch tries to ensure that types are preferentially registered in
the canonical qRegisterMetaType<T>(#T) form.

Where this is not possible, we first check whether the type is already
registered, and only then attempt aliasing to void*/QObject*.

Hid away IsPointerToTypeDerivedFromQObject monstruosity, as QtCore now
provides a macro for this purpose.

Change-Id: I564b6ad3cd14daed9452d85a775c980f5772c703
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>