Fix embedding ActiveQt servers based on native widgets like QGLWidget
Embedding a QGLWidget based server caused QGLWidget to be shown on
a separate toplevel widget. This happened because it is a native widget
and therefore creates a platform window before it is shown, so
_q_embedded_native_parent_handle parameter we use to indicate embedded
windows got set too late and consequently QGLWidget's window didn't
know it was supposed to be embedded.

Fixed by setting the parent in WM_SHOWWINDOW handler in these cases,
and setting the window frameless, which makes platform plugin
re-determine correct native flags for the window, now also taking the
fact that the widget is embedded into account.

Task-number: QTBUG-26438
Change-Id: Id159d864376f773775114c8f00088ce90186b65a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>