Fix crash when handling WM_PAINT during Invoke
Synchronous expose corrupts painter state if it is done during existing
paint operation, which can happen when COM method is invoked in
a signal handler.

Fixed by posting a special message into Windows message queue if there
is a  WM_PAINT in the queue when Invoke is called. This allows Windows
QPA plugin to do an asynchronous expose instead of synchronous one.

Task-number: QTBUG-27209
Change-Id: I93595c1d1dc01601f7a8c4562a145bebb29268b6