Re-enable and test the ActiveQt CMake files.
The Qt5AxContainer CMake module is the most commonly useful, and
the more simple.

Using Qt5AxServer requires the user to do some post-build actions
currently. In the future, it may be possible to add CMake macros
for that kind of thing, but as most people won't use CMake for
this at all, it is not so important at this time.

 http://thread.gmane.org/gmane.comp.programming.tools.cmake.user/44824

Change-Id: If33938f829fbf53190aee9d4652516e1c450f05c