WIP: Re-enable and test the ActiveQt CMake files.
The tests pass, but as I don't have any experience with ActiveQt,
I'm not really sure how to build stuff with it.

 http://lists.trolltech.com/qt-interest/2004-05/thread00936-0.html
 http://www.itk.org/pipermail/insight-users/2004-May/008416.html

It appears that an RC_FILE equivalent is not needed in CMake, but
I'm not clear on whether the executable still needs to be invoked
with -dumpidl etc.

WIP because waiting for someone who has some clue about activeqt
and cmake to come along.

Change-Id: If33938f829fbf53190aee9d4652516e1c450f05c