dumpcpp: properly clean up temporary metaobjects
All qax_* functions that return QMetaObject instances actually return
instances of QAxMetaObject. To make sure that ~QAxMetaObject is
called, we introduce the qax_deleteMetaObject function and call it in
the appropriate places.

The explicit call to QByteArray::detach is needed because the
affected QMetaObject/Method calls return a QByteArray that directly
holds a pointer to the QByteArrayData contained in QAxMetaObject.
Once the QAxMetaObject is destroyed, the QByteArray objects will hold
pointers to freed memory. Therefore we make a deep copy of the
QByteArray by detaching it.

Task-number: QTBUG-28325

Change-Id: I97613f12f1a30608dd5acd67d0faebb3f39cdf70
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>