Suppress registration of the server examples in CI builds.
It does not make sense to register the server examples on the CI
since the registration step is slow and bound to fail if for example
the QPA plugin fails to load in headless environments.

Detect the CI environment by checking on the environment variable
JENKINS_HOME and turn off the axserver post link step accordingly.

Change-Id: I59e8170f071504ee2b141b8a38b594d5bac734c6
Reviewed-by: Andy Shaw <andy.shaw@digia.com>