Fix broken openglax by introducing QOpenGLFunctions_1_1.
The openglax example relies on deprecated OpenGL 1.1 functions
that are no longer present in OpenGL ES 2. This leads to a
missing symbols linker error. The problem is fixed by inheriting
from QOpenGLFunctions_1_1 that provides implementations of the
missing OpenGL functions.

The GLBox destructor is also hardened with selective delete to
avoid crashing if initializeGL is not called (occurs when
dumping IDL).

Change-Id: If35539749ab76293f98a5d75944f0f393f4d2a3c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>