Doc: replaced \target with \keyword at top of page
A \target whose purpose is to link to the top of a
page (and not to a section within a page) works better
as a \keyword, because \target generates a
new html anchor which, in this case, is not tied to
any title element on the page.
A \keyword links to the page itself, as expected.

Change-Id: I7ce0283f7371de191d3e1b35874b3824bf6d2566
Task-number: QTBUG-48482
Reviewed-by: Venugopal Shivashankar <venugopal.shivashankar@digia.com>