IDC: Add folder containing DLL to search path for dependencies.
IDC crashes when attempting to register a ActiveQt COM DLL as
post-build event, if the DLL depends on other DLL(s) in the same
folder. The reason for this is because IDC loads the DLL using
LoadLibrary, which does NOT add the folder containing the DLL to
the search path when loading dependencies.

This patch fixes this problem by introducing the
LOAD_LIBRARY_SEARCH_DLL_LOAD_DIR flag to LoadLibraryEx and
expanding the input path to become fully qualified.

Change-Id: If41609d48842567984013b1f3f2b00b0a93d5960
Task-number: QTBUG-48586
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>