Add VBScript example for the "hierarchyax" project.
The script demonstrates how to navigate between custom Qt objects
from a scripting language. None of the existing scripts demonstrate
this capability, so the script should therefore be useful.

The background for sending this patch is that scripting ActiveQt
can be a bit tricky. Adding more sample scripts will hopefully
make it easier for newcomers to get started.

Change-Id: I06820339bc92b556229af101ca2cc177da96ccc2
Reviewed-by: Andy Shaw <andy.shaw@theqtcompany.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>