Change "argc" variable to have static lifetime.
Required, since the QApplication object contains a reference back to "argc".
The lifetime of the QApplication object exceeds "argc" on the stack. This in
turn causes QApplication to see a corrupted "argc" values later in the function.

This fixes a nasty crash during IDC-based IDL generation & DLL registration
that was only observed in release build post-build events with VS2012.

Change-Id: Ie8e404060deccadc75560748ad7071b41425cf79
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>