Rewrite qaxserver_main.cpp to use CommandLineToArgvW()
Rewrite the command line processing to use QStrings constructed
from the return value of CommandLineToArgvW().
This ensures that arguments with blanks are processed correctly
and fixes the invocation of .exe targets with spaces via idc
using the -dumpidl option and removes the dependency on
the qWinMain() function exported from QtCore.

Task-number: QTBUG-55332
Change-Id: I2d6cc11f2bff30027664707d03a75650d724a4f9
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>