Fix previous optimization to prevent a crash
In 2faaec4e061c7f531624cad3cba5921ed3a51b20, optimizations were done to
fix warnings as indicated by the Clang model. However the combining of
the two ifs caused a problem as riid might still be equal to
IID_IPropertyNotifySink and be the same as ciid. Therefore the check has
to be the last one done to account for other possibilities first.

Change-Id: I6a5f38beb723407b4bea8ab58c04ebbd1b900140
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>