Unset widget window's parent when WM_DESTROY is sent to prevent a crash
When a control is activated and deactivated in Powerpoint then it will
cause the ActiveX window hosting the QWidget to be destroyed. In this
case we need to set the parent of this widget's own QWindow's HWND to
be 0 so it does not reference the deleted window.

This reflects the handling in WM_SHOWWINDOW where the parent is set for
the widget's own QWindow's HWND.

Change-Id: Ie32d408782cdf94a98cb13eda49f6ccf73cccd51
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>