Get the right interface type for m_spInPlaceSite
We depend on m_spInPlaceSite being an IOleInPlaceSiteWindowless
type as we call a function only available for that type and not the base
one. Therefore, we need to have a pointer to both interfaces, so we can
use the right one when needing IOleInPlaceSiteWindowless. Since not
all controls provide IOleInPlaceSiteWindowless we need to have both
in case IOleInPlaceSite is available.

Task-number: QTBUG-63903
Change-Id: Ieaf039c169347606aec43f53c8cb29e1317db3b9
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit cacf666f95df9790cebf5d5cdb0609cee257fab3)