Enable DllSurrogate activation when building as DLL
DllSurrogate enables launching of the control in a separate dllhost.exe process,
instead of loading it directly into the container process. This can be used for
failure isolation and/or security sandboxing.

DllSurrogate activation is not enabled by default. It has to be enabled explicitly by
calling QAxWidget::setClassContext with a bitmask that includes CLSCTX_LOCAL_SERVER
but _not_ CLSCTX_INPROC_SERVER before creating the control.

REF: https://docs.microsoft.com/en-us/windows/desktop/com/dllsurrogate
Task-number: QTBUG-70983
Change-Id: I46ce7a3f8469f3841b59d5bfe284b9ba4e39568d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>