Reverse the order of unitialization to avoid .Net exceptions
During initialization, we always set the client-site first, then
set the advise-sink. The .Net-generated runtime-callable-wrapper
object disappears when the client-site is reset, which generates
runtime exceptions (not effecting the user or process) when
Unadvise is called later. Calling Unadvise first fixes this
issue, and makes sure we uninitialize in the inverse order in
which we initialized.

Fixes: QTBUG-74096
Change-Id: I0f42be13730a3484b4970f2b79fbf416fcda1954
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>