Move static methods/properties of QAxWidget/Object to the front
The static methods and properties (control, exception()),
are currently at the back behind the COM properties and
functions starting with captical letters due to the usage of
a QMap to sort them alphabetically.

Extract helpers to add them to the QMetaObjectBuilder
directly without storing them in the maps and add them
first.

Change-Id: I774dc44a2c1579e839c26433d37d78eac1e0beb6
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>