Adapt to QString::utf16() now returning char16_t *
Make the code agnostic of the exact return type. In one case, were
filling a C# IntPtr, drop the cast and use .data(), as IntPtr has a
ctor from void* and the code stores an explicit string length, so we
don't need the NUL-termination that utf16() guarantees.

There are a handful of casts remaining, so write a small helper that
allows to comment on certain aspects of the conversion.

Done-with: Marc Mutz <marc.mutz@kdab.com>
Change-Id: I16227077b31f02fbbefe1a83e4dfbd2d5333a0f7
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>