QAxObject/Widget: Enable overwriting clear() by dumpcpp
Rename virtual QAxObjectInterface::clear() to
resetControl() and change QAxObject/Widget::clear()
to a non-virtual function calling it.

resetControl() is more inline with Qt naming conventions
and making clear() non-virtual prevents clashes
when a control generated by dumpcpp has a clear()
method. It is not a general solution for the issue;
but helps this case.

Fixes: QTBUG-83735
Change-Id: I6aa01db6e95477e0407ba40fbc14f4991edf5f3e
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>