Fix mingw warnings
GetProcAddr returns a FARPROC (function pointer with no arguments) for
mingw. In reality GetProcAddr does not result in a FARPROC though so
that we have to cast it to void* before casting it to the "real"
function pointer to tell mingw that we know what we are doing.

Additionally the return value of lastIndexOf was ignored in one place.
Just use it instead of relying on the match having been made.

Change-Id: I007f4732ff563aa01d2140ed8f92d1fc78f1a71b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>