Update commercial license headers
Update header.COMM to the files in tqtc-qtactiveqt.
Examples, documentation files or tests are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4156
Change-Id: Icb54effee172b88f9d7fd71caed7b8bc4e630eca
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>
Reviewed-by: Tuukka Turunen <tuukka.turunen@qt.io>
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>