Implement CMake functions to work with IDC
In qmake we have an implicit execution of IDC in case the axserver
config is enabled. This quite difficult to implement using CMake.
This adds the following CMake functions to work with IDC:
 - The qt6_add_axserver_executable function adds axserver executable,
   generates IDL file and link the produced .tbl to the executable.
 - The qt6_add_axserver_library function adds axserver library,
   generates IDL file and link the produced .tbl to the dll.
 - The qt6_target_idl generates IDL file and link the produced .tbl
   to the executable for the existing target.

This also introduces the QT_SKIP_AX_SERVER_REGISTRATION variable,
which disables global registration of ActiveX servers.

Change-Id: I5695db9d67dac07747894fd4bb37c3e6db408035
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>