Update commercial license headers
Updated header.COMM to the files in tqtc-qtactiveqt.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4936
Change-Id: I7299ef37dca807424752caf52d13e6033eb66aff
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Antti Kokko <antti.kokko@qt.io>