Postprocess moc output to fully qualify types
If a property is declared with a type that is also the name of a member
function, then the generated code will result in conflicts and compile
errors. For instance, a property "AddressEntry" might be of type
"AddressEntry*" and have a getter "AddressEntry", where the type lives
in the same namespace as the current class. This is reproducible with
the qutlook example, which doesn't build.

This issue can also be provoked by valid C++ code, but since moc has
no information about which types are classes or enums, or in which
namespace they live in, we cannot do anything about that problem (yet).

However, dumpcpp has information about which namespace a type live in,
so we can postprocess the moc output and fix the problematic code by
fully qualifying the types used with the namespace they were generated
in.

This slows down dumpcpp runs quite a bit, but those are infrequent as
the type library typically doesn't change.

Fixes: QTBUG-100145
Pick-to: 6.4 6.3
Change-Id: Id9b4656cb1ff2c319e0b87bd22b7e9399e7c410d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>