Use QObject::metaObject instead of QAxBase::axBaseMetaObject in qaxbase.cpp
Using QAxBase::axBaseMetaObject results in usage of the dynamically
queried COM object whenever it is used. This behavior is not always what
we want though. If a type library was queried for its interface with
dumpcpp we should use the information from the resulting h and cpp files
and do not operate on the information given by the COM object. The files
generated by dumpcpp will override the virtual metaObject function.

By using QObject::metaObject instead of QAxBase::axBaseMetaObject we
make sure that:
a) Dynamically queried information is used if no overload has been added
by dumpcpp. This is the case because QAxWidget::metaObject and
QAxObject::metaObject return QAxBase::axBaseMetaObject. So the
functionality there does not change.
b) If dumpcpp was run on a type library the types in the code generated
by dumpcpp will overload QObject::metaObject. In this cases we will only
use the information from dumpcpp runs.

With the previous approach it was possible that we run into issues if
there are inconsistencies between a user's type library and the actual
COM interface. If a function was available in the COM interface but not
inside the type library definitions that might have thrown off ActiveQt
which resulted in calls of wrong versions. This regression (it worked as
expected in Qt5) is fixed by this patch.

Pick-to: 6.4 6.3 6.2
Change-Id: I1af5ecb81f0c4ce0dc1e899cc3f019ebee575f4b
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>