Disable headersclean check for the ActiveQt module
We disable this headersclean check this way, since syncqt adds the
header files from the subdirectories to the ActiveQt module. This
triggers the headersclean rules creation and execution, but they
fail because ActiveQt module doesn't link the necessary libraries so
their compile flags and include directories are missing.

We cannot simply link libraries because the missing include directories
belong to the AxServer and AxContainer modules which have conflicting
object files and cannot be linked in the same time.

Change-Id: I7b8cc40dc920d1f2d436bededb76c8ff270bbc6d
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>