Fix returning QAxServer objects from methods
A QVariant initialized from a QMetaType is valid, but null, so in case
of pointer return types we didn't update the QVariant that would then be
written back into the return value.

Always write back if we have previously recognized that we are dealing
with a pointer return type, indicated by argv[0] pointing at the
argv_pointer array.

Task-number: QTBUG-106024
Pick-to: 6.4
Change-Id: Ic1bd258b006746e0233c09fe68842bdeefdfc84f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>