CMake: Fix "no rc file" case when running idc
In case if no rc file is present in the sources of a target that
has qt6_add_axserver_library called on it, the build would fail
with unknown command when trying to run the "echo" part of the
command.

This happens because cmd.exe interprets the echo and its argument as a
single argument, despite the space between then.

Use $<SEMICOLON> in conjunction with COMMAND_EXPAND_LISTS to ensure
the echo and the argument that follows it are split.
Also remove the extra quotes which are not needed.

Tested to work with powershell and cmd.exe with
CMake 3.16.9, 3.17.4, 3.18.4, 3.21.2, 3.24.2.

Amends 2b8681267a1f79734b379b7d7776c520de7be5b5

Pick-to: 6.4
Change-Id: If81f7f6947e35709ce4cf1a0333247fb238cf73a
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>