Fix invalid QApplication construction
QApplication construction with argc=0 is incorrect according to the doc.
The construction pattern is furthermore causing problems in QtWebEngine,
which is expecting the executable name as first command-line arguments
as reported in QTBUG-110157.

Fix problem by using the Microsoft-specific __argc and __argv globals.
This ensures that the actual command-line arguments are used for
QApplication construction.

Pick-to: 6.5
Task-number: QTBUG-110158
Change-Id: Ib8e6ac7a1c24cfabfdafe8489fe8b8c4765753fd
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>