Fix QAxScriptManager::call
It takes the name of the function, not of the script name so use the
scriptForFunction helper to find the QAxScript object that provides
access to the function.

The previous code implies that this should work for both function names
and for full function prototypes, so implement that correctly and
without allocating too many unnecessary temporary objects.

Update the test that is now passing to cover all scenarios.

Change-Id: I939942cc805a56212d70b1a12b95aaa4b2f64d59
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 855472a28675f11778a54ef1e42f6114d924b9e2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>