Add CMakeUserPresets.json to .gitignore in ActiveQt
The CMakeUserPresets.json file can be used to simplify configuring Qt,
particularly with tools such as Visual Studio and Visual Studio Code.

As opposed to the CMakePresets.json, the CMakeUserPresets.json should
only contain user specific configurations, and should never be checked
into source code management system. This has already been added to the
top level repo but to ease working with ActiveQt as a module, it should
also be included here.

Pick-to: 6.5 6.6
Change-Id: I1c0139711aab414dfd07b34623d9356ede523bc3
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>