Block signals while the meta object gets created
If a COM object emits events and calls QAxEventSink::Invoke while the
metaobject is getting constructed, then we might end up with accessing
incomplete initialized meta object data when mapping the Invoke call to
the corresponding Qt signal.

Use a QSignalBlocker while we are generating the meta object, so that
the QAxEventSink::Invoke and OnChanged implementations exit early.

Pick-to: 6.6 6.5 5.15
Fixes: QTBUG-100657
Fixes: QTBUG-96871
Change-Id: If9368ee878bbe56e2fb61ce251383841191b3a6d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Wladimir Leuschner <wladimir.leuschner@qt.io>