Describe how to deal with COM reentrancy problems in ActiveQt
If two COM objects in two single threaded apartments can communicate and
both objects can initiate a call (typical situation if both client and
server have an user interface), the implementations must deal with
reentrancy problems. This is not always an obvious property of out-of-
process COM servers, and documenting it can help developers to solve or
prevent such issues.

Pick-to: 6.5 6.6
Change-Id: Ibb1bb7e7a58709fdb23feb3ae68c9cbf62200d4b
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>