Remove the SimpleQML example
It's undocumented, and it uses outdated QML idioms such as setting a
context property when it could use a singleton.
Also, the extra QMainWindow + Controller class seems unnecessary; a
QQuickWidget subclass that provides the "Controller" functionality would
work as well. As it is, the "Controller" API isn't accessible for COM
clients.

Change-Id: I42d8e00fb13311e8b5df74f69f9bb76053346229
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>