Run dumpcpp also when type library is specified by UUID
We should only give a value to the custom command's MAIN_DEPENDENCY if
the type library is a file, not when it is an UUID. If the
MAIN_DEPENDENCY does not exist as a target or a file, the custom command
is not run at build time.

This regression was introduced in
377c95c4a09f7c363c5c609c2bf7907373c39925 when the MAIN_DEPENDENCY was
introduced to reevaluate the custom command if the type library file
changed.

Fixes: QTBUG-126518
Change-Id: Idc5346049ce9e5215827a454c3031e5062fb010d
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
(cherry picked from commit 24ea2efe32d6f78e84e0f2741cea1b542c20f00c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>