QAxServer: Do not trigger sendPostedEvents by WM_QT_SENDPOSTEDEVENTS
If we trigger sendPostedEvents when we receive a WM_QT_SENDPOSTEDEVENTS
message in our custom hook in a QAxServer we can run into an endless
loop of receiving this event and triggering of exactly the same.
This is due to changes in the event handling in qtbase. We should just
ignore this event type in our custom hook.

The hook will be completely removed in a followup patch. This
intermediate step is just done to be able to partially revert the patch
in lower branches in case of breakages.

Pick-to: 6.8 6.5
Fixes: QTBUG-123498
Change-Id: I3e817dc03b1acbe76d8dc00fefaa7f7f88205237
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>