QAxServer: Do not trigger sendPostedEvents by WM_QT_SENDPOSTEDEVENTS
If we trigger sendPostedEvents when we receive a WM_QT_SENDPOSTEDEVENTS
message in our custom hook in a QAxServer we can run into an endless
loop of receiving this event and triggering of exactly the same.
This is due to changes in the event handling in qtbase. We should just
ignore this event type in our custom hook.

Pick-to: 6.5
Fixes: QTBUG-123498
Change-Id: I3e817dc03b1acbe76d8dc00fefaa7f7f88205237
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>