Make QAxBase/QAxWidget QDataStream operators hidden friends
Prompted by 4664acf185b056e38933bd42a29bc012f1011d0b adding QAxWidget
operators as non-member non-friends, make both QAxBase and QAxWidget
operator hidden friends, as we want all our operators to be these days.

This change is BC, but it's not 100% SC, because streaming of objects
that merely implicitly convert to QAxWidget/QAxBase, but are not
derived from them, now fails.

Also move the QAxWidget operators out-of-line, so we have some chance
of adding something to them, though the pre-existing QAxBase operators
mean adding anything there will change the serialization format
vis-a-vis 6.8 (nothing a QDataStream version upgrade can't solve,
though).

[ChangeLog][Potentially Source-Incompatiable Changes][QAxBase] The
QDataStream operators are now hidden friends, so they will only match
types which are QAxBase, or derived from it, not those that merely
implicitly convert to them. A backwards-compatible fix is to make the
conversion explicit and only call these operators on actual QAxBase
subclass objects.

Amends the start of the public history and
4664acf185b056e38933bd42a29bc012f1011d0b.

Pick-to: 6.9
Change-Id: I60fe1ba10f7478c2dfedf33cc45bbc257f297a4b
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>