dumpcpp: Qualify type names
Changes in Qt 6.9's moc lead to compilation errors in dumpcpp's C++
output. There are two separate issues:

1. Type libraries may contain properties/methods that have the same name
as a type that's defined in the TLB. That led to a type vs member
conflict in moc's output. That's resolved in moc itself.

2. In moc's output, types that are declared in the TLB must be qualified
with the TLB's namespace to resolve name conflicts.

This patch does the latter:
When writing the in-memory header file that's passed to moc we qualify
types that belong to the type library with the type library's namespace.

Pick-to: 6.9 6.10
Fixes: QTBUG-136512
Fixes: QTBUG-137347
Change-Id: I27858f145261f0aee392bef46756691ae0360dd2
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>