dumpcpp: Fix assert when hitting unknown parameter types in type libraries
It is possible that type libraries reference types from other type
libraries. If these dependencies were unknown dumpcpp ran into an assert
which resulted in incomplete .h and .cpp files being generated.

As this is a regression from Qt 5 and it is impossible to obtain this
type information without further information about additional tlb files
we just skip these functions. They will not be available to users but
the rest of the interface is usable.

Task-number: QTBUG-140884
Pick-to: 6.10 6.8
Change-Id: I66a1006718c40519b4a7e9a2b3596afffef6a9d0
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>